+++ src/components/Login/index.tsx
@@ -1,0 +1,2 @@
+export { default as LoginTitle } from './LoginTitle';
+export { default as LoginForm } from './LoginForm';
